mcp_server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
mysql-connector-python>=9.0.0
pymssql>=2.2.0
pydantic>=2.0.0
mcp>=1.0.0
anthropic>=0.42.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pdfplumber>=0.11.0
